straightforward motion of the Gaussian algorithm long double -> double

+++ LinearAlgebra/Matrix.h
@@ -6,7 +6,7 @@ using std::vector;
 class Matrix
 {
   int n, m;
-  vector<vector<long double>> matrix;
+  vector<vector<double>> matrix;
 
 public:
   Matrix();
@@ -16,7 +16,7 @@ public:
 
   const Matrix& operator=(const Matrix &copy);
   
-  vector<long double> GaussianElimination(vector<long double> terms);
+  vector<double> GaussianElimination(vector<double> terms);
   void swapRows(const int firstRow, const int secondRow);
   void swapColumns(const int firstCol, const int secondCol);
